Removing a file that was renamed by the last commit
