+++ Gems/Multiplayer/Code/Source/Components/SimplePlayerSpawnerComponent.cpp
@@ -8,36 +8,36 @@
 #include "AzCore/Component/TransformBus.h"
 
 #include <Multiplayer/IMultiplayer.h>
-#include <Multiplayer/Components/SimpleNetworkPlayerSpawnerComponent.h>
+#include <Multiplayer/Components/SimplePlayerSpawnerComponent.h>
 #include <AzCore/Console/ILogger.h>
 #include <AzCore/Serialization/SerializeContext.h>
 #include <AzCore/Serialization/EditContext.h>
 
 namespace Multiplayer
 {
-    void SimpleNetworkPlayerSpawnerComponent::Activate()
+    void SimplePlayerSpawnerComponent::Activate()
     {
         AZ::Interface<IMultiplayerSpawner>::Register(this);
     }
 
-    void SimpleNetworkPlayerSpawnerComponent::Deactivate()
+    void SimplePlayerSpawnerComponent::Deactivate()
     {
         AZ::Interface<IMultiplayerSpawner>::Unregister(this);
     }
 
-    void SimpleNetworkPlayerSpawnerComponent::Reflect(AZ::ReflectContext* context)
+    void SimplePlayerSpawnerComponent::Reflect(AZ::ReflectContext* context)
     {
         if (const auto serializeContext = azrtti_cast<AZ::SerializeContext*>(context))
         {
-            serializeContext->Class<SimpleNetworkPlayerSpawnerComponent, AZ::Component>()
+            serializeContext->Class<SimplePlayerSpawnerComponent, AZ::Component>()
                 ->Version(1)
-                ->Field("PlayerSpawnable", &SimpleNetworkPlayerSpawnerComponent::m_playerSpawnable)
-                ->Field("SpawnPoints", &SimpleNetworkPlayerSpawnerComponent::m_spawnPoints)
+                ->Field("PlayerSpawnable", &SimplePlayerSpawnerComponent::m_playerSpawnable)
+                ->Field("SpawnPoints", &SimplePlayerSpawnerComponent::m_spawnPoints)
                 ;
 
             if (AZ::EditContext* editContext = serializeContext->GetEditContext())
             {
-                editContext->Class<SimpleNetworkPlayerSpawnerComponent>(
+                editContext->Class<SimplePlayerSpawnerComponent>(
                                "Simple Network Player Spawner",
                                "A simple player spawner that comes included with the Multiplayer gem. Attach this component to any level's root entity which needs to spawn a network player."
                                         "If no spawn points are provided the network players will be spawned at the world-space origin.")
@@ -47,12 +47,12 @@ namespace Multiplayer
                            ->Attribute(AZ::Edit::Attributes::AppearsInAddComponentMenu, AZ_CRC_CE("Level"))
                            ->DataElement(
                                AZ::Edit::UIHandlers::Default,
-                               &SimpleNetworkPlayerSpawnerComponent::m_playerSpawnable,
+                               &SimplePlayerSpawnerComponent::m_playerSpawnable,
                                "Player Spawnable Asset",
                                "The network player spawnable asset which will be spawned for each player that joins.")
                            ->DataElement(
                                AZ::Edit::UIHandlers::Default,
-                               &SimpleNetworkPlayerSpawnerComponent::m_spawnPoints,
+                               &SimplePlayerSpawnerComponent::m_spawnPoints,
                                "Spawn Points",
                                "Networked players will spawn at the spawn point locations in order. If there are more players than spawn points, the new players will round-robin back starting with the first spawn point.")
                 ;
@@ -60,17 +60,17 @@ namespace Multiplayer
         }
     }
 
-    void SimpleNetworkPlayerSpawnerComponent::GetProvidedServices(AZ::ComponentDescriptor::DependencyArrayType& provided)
+    void SimplePlayerSpawnerComponent::GetProvidedServices(AZ::ComponentDescriptor::DependencyArrayType& provided)
     {
         provided.push_back(AZ_CRC_CE("MultiplayerSpawnerService"));
     }
 
-    void SimpleNetworkPlayerSpawnerComponent::GetIncompatibleServices([[maybe_unused]] AZ::ComponentDescriptor::DependencyArrayType& incompatible)
+    void SimplePlayerSpawnerComponent::GetIncompatibleServices([[maybe_unused]] AZ::ComponentDescriptor::DependencyArrayType& incompatible)
     {
         incompatible.push_back(AZ_CRC_CE("MultiplayerSpawnerService"));
     }
 
-    NetworkEntityHandle SimpleNetworkPlayerSpawnerComponent::OnPlayerJoin(
+    NetworkEntityHandle SimplePlayerSpawnerComponent::OnPlayerJoin(
         [[maybe_unused]] uint64_t userId, [[maybe_unused]] const MultiplayerAgentDatum& agentDatum)
     {
         const PrefabEntityId prefabEntityId(AZ::Name(m_playerSpawnable.m_spawnableAsset.GetHint().c_str()));
@@ -78,7 +78,7 @@ namespace Multiplayer
 
         if (m_spawnPoints.empty())
         {
-            AZLOG_WARN("SimpleNetworkPlayerSpawnerComponent is missing spawn points. Spawning new player at the origin.")
+            AZLOG_WARN("SimplePlayerSpawnerComponent is missing spawn points. Spawning new player at the origin.")
         }
         else
         {
@@ -116,8 +116,8 @@ namespace Multiplayer
         return controlledEntity;
     }
 
-    void SimpleNetworkPlayerSpawnerComponent::OnPlayerLeave(ConstNetworkEntityHandle entityHandle, [[maybe_unused]] const ReplicationSet& replicationSet, [[maybe_unused]] AzNetworking::DisconnectReason reason)
+    void SimplePlayerSpawnerComponent::OnPlayerLeave(ConstNetworkEntityHandle entityHandle, [[maybe_unused]] const ReplicationSet& replicationSet, [[maybe_unused]] AzNetworking::DisconnectReason reason)
     {
         AZ::Interface<IMultiplayer>::Get()->GetNetworkEntityManager()->MarkForRemoval(entityHandle);
     }
-} // namespace SimpleNetworkPlayerSpawnerComponent
+} // namespace SimplePlayerSpawnerComponent
